refactor: use hooks to refactor form
